Schema dumper refactoring: processing stored into different module
